requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
gitdb2==2.0.3
GitPython==2.1.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
requests==2.18.4
smmap2==2.0.3
urllib3==1.22
Werkzeug==0.14.1
